+++ client/src/app/pages/applications/application-review/components/application-assessment-summary-table/application-assessment-summary-table.tsx
@@ -25,7 +25,7 @@ import {
   FilterToolbar,
   FilterType,
 } from "@app/shared/components/FilterToolbar/FilterToolbar";
-import { useFilterState } from "@app/shared/hooks/useFilterState";
+import { useLegacyFilterState } from "@app/shared/hooks/useLegacyFilterState";
 
 interface ITableItem {
   answerValue: string;
@@ -90,7 +90,7 @@ export const ApplicationAssessmentSummaryTable: React.FC<
     },
   ];
 
-  const { filterValues, setFilterValues, filteredItems } = useFilterState(
+  const { filterValues, setFilterValues, filteredItems } = useLegacyFilterState(
     tableItems || [],
     filterCategories
   );
