+++ client/src/app/pages/external/jira/trackers.tsx
@@ -36,7 +36,7 @@ import {
   Th,
   TableComposable,
 } from "@patternfly/react-table";
-import { useTableControls } from "@app/shared/hooks/use-table-controls";
+import { useLocalTableControls } from "@app/shared/hooks/table-controls";
 import { SimplePagination } from "@app/shared/components/simple-pagination";
 import {
   ConditionalTableBody,
@@ -86,7 +86,7 @@ export const JiraTrackers: React.FC = () => {
     onDeleteInstanceError
   );
 
-  const tableControls = useTableControls({
+  const tableControls = useLocalTableControls({
     items: jiraTrackers,
     columnNames: {
       name: "Instance name",
