+++ client/src/app/pages/identities/identities.tsx
@@ -28,7 +28,7 @@ import {
   NoDataEmptyState,
 } from "@app/shared/components";
 import { Identity, ITypeOptions } from "@app/api/models";
-import { useFilterState } from "@app/shared/hooks/useFilterState";
+import { useLegacyFilterState } from "@app/shared/hooks/useLegacyFilterState";
 import { useLegacyPaginationState } from "@app/shared/hooks/useLegacyPaginationState";
 import { useLegacySortState } from "@app/shared/hooks/useLegacySortState";
 import { useEntityModal } from "@app/shared/hooks/useEntityModal";
@@ -134,7 +134,7 @@ export const Identities: React.FC = () => {
     },
   ];
 
-  const { filterValues, setFilterValues, filteredItems } = useFilterState(
+  const { filterValues, setFilterValues, filteredItems } = useLegacyFilterState(
     identities || [],
     filterCategories
   );
