+++ client/src/app/pages/waves/wave-applications-table/wave-applications-table.tsx
@@ -17,7 +17,7 @@ import {
   Tr,
 } from "@patternfly/react-table";
 import alignment from "@patternfly/react-styles/css/utilities/Alignment/alignment";
-import { useTableControls } from "@app/shared/hooks/use-table-controls";
+import { useLocalTableControls } from "@app/shared/hooks/table-controls";
 import {
   ConditionalTableBody,
   TableHeaderContentWithControls,
@@ -36,7 +36,7 @@ export const WaveApplicationsTable: React.FC<IWaveApplicationsTableProps> = ({
 }) => {
   const { t } = useTranslation();
 
-  const tableControls = useTableControls({
+  const tableControls = useLocalTableControls({
     items: applications,
     columnNames: {
       appName: "Name",
