--- client/src/app/shared/hooks/use-table-controls/pagination/index.ts
@@ -1,2 +1,0 @@
-export * from "./usePaginationState";
-export * from "./usePaginationDerivedState";
