--- client/src/app/shared/hooks/use-table-controls/sorting/index.ts
@@ -1,2 +1,0 @@
-export * from "./useSortState";
-export * from "./useSortDerivedState";
