+++ client/src/app/shared/hooks/useLegacySortState.ts
@@ -3,9 +3,9 @@ import { ISortBy, SortByDirection } from "@patternfly/react-table";
 import i18n from "@app/i18n";
 
 // NOTE: This was refactored to expose an API based on columnKey instead of column index,
-//       and to return generic state data and decouple the PF props piece to another helper function.
+//       and to return generic state data and decouple the client-side-sorting piece to another helper function.
 //       See useSortState for the new version, which should probably be used instead of this everywhere eventually.
-//       See useSortProps for the prop parts that were removed here (TODO ?)
+//       See useLocalSortDerivedState and useTableControlProps for the parts that were removed here.
 
 export interface ILegacySortStateHook<T> {
   sortBy: ISortBy;
